fixup for Equip/ fixed, add rcm.c rifle.c tabako.c aam.c kogaku2.c bomb.c mine.c
