--- src/Equip/kogaku2_act_nullsub_800615F4.c
@@ -1,8 +1,0 @@
-#include "actor.h"
-
-extern void EQ_InvisibleUnit_80060E68(int param_1, int param_2, int param_3);
-
-void kogaku2_act_nullsub_800615F4(void)
-{
-
-}
