--- src/Equip/kogaku2_kill_80061508.c
@@ -1,10 +1,0 @@
-
-
-extern void kogaku2_kill_helper_80061384(int param_1);
-
-void kogaku2_kill_80061508(int param_1)
-
-{
-    kogaku2_kill_helper_80061384(param_1);
-    return;
-}
